CMake: work around an issue with CMake Xcode generator and the Swift driver
The Swift driver does not update the mtime of the output files if the
existing output files on disk are identical to the ones that are about
to be written.  This behavior confuses the makefiles used in CMake Xcode
projects: the makefiles will not consider everything up to date after
invoking the compiler.  As a result, the standard library gets rebuilt
multiple times during a single build.

To work around this issue we touch the output files so that their mtime
always gets updated.