[Type checker] Don't fold the parentheses around call arguments into a TypeExpr.
TypeExpr formation from expressions is eager, and was pulling in the
parentheses used to form a call argument, e.g.,

  strideof(Int)

would end up being a CallExpr whose argument was a TypeExpr containing
the type "(Int)". This had the side effect of not requiring ".self"
where it should (as reported in SR-899), e.g., the correct code for
the above would be:

  strideof(Int.self)

Don't pull the parentheses of a call argument into the TypeExpr, so we
get a more sane AST. This would fix SR-899, but because SE-0090 is in
limbo right now, we don't actually want to require the initial code to
be considered well-formed right now. Therefore, put a hack into the
checking for ".self" that ignores this specific case.