[Type checker] Be more rigorous about extracting argument labels from calls.
Whenever we have a call, retrieve the argument labels from the
argument structurally and associate them with the callee. We were
previously doing this as a separate AST walk (which was unnecessary),
so fold that into constraint generation for a CallExpr.

This is a slightly-pared-back version of
3753d779bc13d13b2a2d9ccf828938938fb7f3a1 that isn't so rigid in its
interpretation of ASTs. I'll tighten up the semantics over time.