Fix Float/Double .nextUp/.nextDown on arm(v7).
On armv7, leastNonzeroMagnitude is defined to be leastNormalMagnitude.
This seems wrong to me from a language point of view, because it's
possible to have a nonzero float that compares less than the leastNonzero float.
However, it must have been done to gloss over hardware rounding errors.
This is not a problem on arm64.

As a result, nextUp/Down methods were not self-consistent. I fixed
this by skipping over subnormal representations, following the
precedent set by leastNonZeroMagnitude.

This fixes the FloatingPoint floatNextUpDownTests on armv7.