Sema: Fix for derived classes where the base class has inaccessible initializers
If the base class is in a different module and defines an internal
initializer, or it is in a different file and defines a private
initializer, the subclass cannot access this initializer.

Fix this by insisting on synthesizing initializers even if they
override an internal or private initializer in the base class.
In this case, synthesize them as a stub instead of a chaining
initializer. They cannot be called, and they cannot reference
the superclass initializer, so a stub is appropriate here.

Also in SILGen, emit vtable entries for stub initializers unless
they are overriding an Objective-C initializer.