Integrate fileprivate
The majority of changes here are just plumbing needed to get the
compiler to look at the (new!) FilePrivate as the (old!) Private and
the (new!) Private as an even more closed-off form the of (old!)
Private.

The interesting changes are in `lib/AST/NameLookup.cpp` where (new!)
private is defined as a declaration where our declcontext and their
declcontext have to match or nest.  (new!) Private predictably causes
quite a lot more dead code elimination to fire in places which has lead
to some interesting bugs that I will see about patches for (or at least
SRs).