XFAIL a compiler_crashers regression
This now fails with the infamous 'Assertion failed: ((conforms ||
replacement->is<ErrorType>() || firstArchetype->getIsRecursive() ||
isOpenedAnyObject(replacement) || replacement->is<GenericTypeParamType>())
&& "Constraint system missed a conformance?")' error.

The actual problem is something with unbound generics and inference
of generic parameters. Here's a minimal test case:

protocol A {}

struct B<f> {}

func a<T:A>() {
  class AA {
    var t:B=B()
  }
}

I'll fix it soon.