Sema: Fix memory corruption in another circularity check
There was a potential memory corruption issue if the decl was
already determined to be part of a cycle by the time we call
this function. In this case, we would walk off the end of
an empty vector.

Start with a non-empty path vector and assert that it is
non-empty to avoid this.

I don't have a test for this either, again it fixes regressions
from a future patch here.