Sema: Use PointerEmbeddedInt instead of Fixnum (#3009)
PointerEmbeddedInt is provided by stock LLVM and serves the same purpose.  This
updates the coersion function to use that rather than the swift specific Fixnum.
Furthermore, this type has been removed in the future in favour of the
PointerEmbeddedInt type, which makes this change something which will be needed
when rebasing to a newer LLVM release.