[update-checkout] Rename the 'branch' concept to 'scheme'.
update-checkout as a script has two different concepts associated with the term
'branch':

1. A "real" git branch.
2. A dictionary of names mapped to lists of (repo, git branch) pairs.

The second concept is used to specify on the command line a "palette" of
branches in all of the repositories associated with things like "master",
"master-next", and the preview branches. This is far more than a "branch". This
commit gives a unique name, "scheme" to that concept.

Keep in mind that even though the internals have changed, the outside interface
(i.e. --branch) is still the same.

That will change at a later time since I would like to change the interface
slightly which may require bots to be updated.