runtime: add missing include
The use of `std::make_tuple` requires the declaration of the function which is
provided by the C++ standard header `tuple`.  This would get implicitly included
by one of the other C++ headers on the currently tested libraries, but does not
occur with the Microsoft C++ library.  Add the missing include.