This test regressed after recent changes
We fail to emit a diagnostic here, because name lookup is able to find the
associated type, even though we don't build a valid generic signature in
this case.

Some more changes to name lookup to untangle "find members in protocols"
from "find concrete witness" flags are needed. I'll do this soon, as part
of fixing unqualified lookup of protocol typealiases, which currently
doesn't work either.