Fix archetype access path binding
When considering a type that declares an archetype bound to an
associated type, we were not generating the access paths to any
underlying conformance data and so were not properly able to grab the
witness table for those conformances resulting in a crash.  This way,
all bound requirements also make sure to bind archetype access paths so
we can see actually see the conformances.