[cmake] Do not use LLVM's configure_lit_site_cfg to configure out lit config.
We do not use any variables in our lit.site.cfg.in files that are not already
set at this point in the configuration. So all calling this function does is
create some variables that are not used and then calls configure.

This commit just removes the call to the function and calls configure directly.
This simplifies the code (we are not calling into llvm) and reduces exposure to
llvm internals changing.

rdar://26154980