build: improve static library generation
Modern `ar` implementations support constructing the index table on the
construction.  This means that we dont need to re-read the contents and run a
second program after the initial library construction to get current behaviour.
This is a simplification of the generated build rules which improves performance
since the data is not re-read nor a second tool invoked.