stubs: export symbols properly
These two symbols were not correctly scoped, placing them inside of the swift
namespace.  This would change the linkage names and fail to link as a result.
Scope one as the header properly places it in an extern "C" block.  The gyb file
is unable to find the desired runtime header, so explicitly mark the exported
symbol as being exposed with C linkage.