[update-checkout] Only use the default-branch-scheme when cloning.
We want to allow for people to use update-checkout to just update branches even
in the case where they are not using a scheme (for instance if they have a side
branch).

This means though that one can still clone and get all of the appropriate
repos/branches.

The only unfortunate thing about this change is that when cloning, already
cloned repositories that already are setup to a branch will have their branch
changed.

It would be good at some point to separate updating from cloning. It would make
update-checkout much cleaner.