[build-script] When building with Xcode, do not build or test any stdlib but the macosx-x86_64 stdlib.
There is a known bug where Xcode can not properly cross compile swift standard
libraries since it always wants to pass in -mmacosx-version-min (see
CMakeLists.txt:512).

Even though this code is in CMake, build-script if it is not told to do the
correct thing, will still try to build those cross compiled libraries. This will
fail since cmake did not create a target.