[test/IRGen] Fix a test failure in runtime_calling_conventions.swift
Fix a mistake introduced by 1409a445, which moved us to the new
setUnnamedAddr API introduced by llvm/r272709.

rt_swift_release is linkonce_odr, so each module gets an identical copy
of it and its address doesn't matter. Set the local_unnamed_addr
attribute on it.

Thanks to Saleem Abdulrasool for the catch!