SILOptimizer: Don't inline functions that directly reference the metatype of $Self
Formerly SILGen would never emit this sequence. In fact in most places
we lower away dynamic Self, replacing it with the concrete Self type
instead.

However, with an upcoming change, I'm using 'metatype $Self' as a handy
way to grab IRGenSILFunction::LocalSelfMetadata, since that's what it
already does.

Note that the tests for this are in the next patch.