[cmake] Create/wireup a new API for linking against imported cmake libraries called LINK_IMPORTED_LIBRARIES.
I am going to use this to handle our dependencies on clang/llvm in a
clean way. I am going to upstream a patch (hopefully) to cmark that will
allow it to use the same functionality as well.

In this patch, we just let through the LINK_IMPORTED_LIBRARIES and we
only use it for clang dependencies (preserving the current behavior). In
subsequent commits, I am going to introduce helper commits that fix the
Xcode release llvm + debug swift build.

rdar://27062396