[cmake] Create/wiretup function swift_config_imported_libraries.
This new function takes in imported library names and produces full
paths to the appropriate build directory for the imported library.

In the Ninja case, this does nothing, but in the Xcode case without
this, a debug swift build will always look for clang flags in the debug
LLVM build directory. This is a problem when we are producing release
LLVM build artifacts.

rdar://27062396