Sema: Fix resolveTypeInContext() for typealiases in protocols
Name lookup won't even find typealiases in protocols or protocol
extensions in some cases, but it does if the type is referenced
from expression context, which exposed a crash.

Fix the crash by generalizing the logic a bit. This will be more
generally useful once I fix a few more issues in name lookup.