IRGen: combine alloc_stack/copy_addr value witness calls.
This optimization inside IRGen combines up to 2 alloc_stack (with optional copy_addr [initialize]) calls into a single swift_stackAllocXx call.
And it eliminates the need of a runtime call for the associated dealloc_stack (therefore it also makes sense to do it for a single alloc_stack).

For details see the comment in Heap.cpp.

The main goal is to reduce code size of generic functions.
With this optimization generic functions can be reduced up to 20%.
The code size of the whole swift stdlib is reduced by 5%.

In addition to that, it improves performance a lot (I've seen up to 8x) if generic types are used which do not fit into the 3-word inline buffer.
This is due to the use of a bump pointer allocation which reduces the number of mallocs significantly.

The benchmark results show a few improvements for -O (up to 20%), although I don't know which are noise.
But there are some significant wins for -Onone (up to 8x) because -Onone mainly calls into generic non-specialized stdlib functions.

This is temporary work: TODO: source code documentation, testing